Extract StageManager into a standalone class
